refactor(executors): extract pure prompt + composer helpers from cursor (1576→1451)
Port of upstream diegosouzapw/OmniRoute#5960. Extracts pure prompt and
composer helpers into open-sse/executors/cursor/{prompt,composer}.ts; the
host imports them back. Adds a split-boundary unit test.
